Add random number generation support for CPU and Metal backends
- Implemented a new random number generation package for both CPU and Metal backends, utilizing the Philox-4×32-10 algorithm for consistent output across platforms.
- Introduced functions for generating standard-normal float32 values, ensuring bitwise reproducibility between CPU and Metal implementations.
- Added Metal kernel and dispatch functions to handle random number generation efficiently on the GPU, with appropriate error handling for Metal context and command buffer management.
- Developed comprehensive tests to validate the correctness and performance of the random number generation across different scenarios, including known-answer tests and ULP tolerance checks.
- Updated documentation to reflect the new random number generation capabilities and usage instructions.

+++ device/cpu/allocate_test.go
@@ -21,7 +21,10 @@ func TestBackendAllocateAligned(t *testing.T) {
 			pointer, err := backend.allocateAligned(128)
 			So(err, ShouldBeNil)
 			So(pointer, ShouldNotBeNil)
-			So(uintptr(pointer)%workspaceAlign, ShouldEqual, 0)
+			// Match types: uintptr(pointer)%workspaceAlign is uintptr.
+			// GoConvey's ShouldEqual is type-strict; passing bare 0 (int)
+			// causes "type difference" failure.
+			So(uintptr(pointer)%workspaceAlign, ShouldEqual, uintptr(0))
 		})
 	})
 }
